Add files via upload
This is the result of a presentation of transactional data containing all transactions that occurred between 01/12/2010 and 09/12/2011 for non-store online retailers based in the UK and registered. The company mainly sells unique gifts for all occasions. Many of the company's customers are wholesalers, with a data amount of more than 5 thousand. LINK: https://archive.ics.uci.edu/dataset/352/online+retail